Support all the maps
Added each of the standard maps.
Added coordinate calibrations for each map that should be good, but are mostly untested.
When a development version is run there are now additional buttons for map calibration and dummy map points.
Loading species in the background is now more visible, with status bar text and showing the loading cog.