Support searching for tamed creatures by name
Added a search box when showing tames. Results are highlighted on the map and in the creature list.

Result list now requires items to be selected before they are highlighted on the map, but you can now select multiple creatures and have them stay highlighted.

ISSUE: With many creatures in the results list, sometimes highlights are not shown immediately. Scrolling the list seems to fix it. The map highlights correctly.